% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.matrix.R
\name{correlation.matrix}
\alias{correlation.matrix}
\title{A shiny app to edit correlation matrix}
\usage{
correlation.matrix(DF, outdir = getwd(), outfilename = "table")
}
\arguments{
\item{DF}{the original dataset that need to put in in order to invoke the app}

\item{outdir}{the file directory}

\item{outfilename}{the output file name}
}
\value{
a dataframe that is the eidited correlation matrix as dataframe format.
}
\description{
A shiny app to edit correlation matrix
}
\examples{
correlation.matrix(DF)

}
