% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corrmatrix.limit.R
\name{Corrmatrix.limit}
\alias{Corrmatrix.limit}
\title{Checking correlations for feasibility}
\usage{
Corrmatrix.limit(start, end, number.of.scores)
}
\arguments{
\item{start}{a number to set up the min value for the discrete variable.}

\item{end}{a number to set up the max value for the discrete variable.}

\item{number.of.scores}{a number represent  how many discrete variables.}
}
\value{
The functions returns a list of two matrices:
the former contains the lower bounds, the latter the upper bounds of the
 feasible pairwise correlations (on the extra-diagonal elements)
}
\description{
The function returns the lower and upper bounds of the
 correlation coefficients of each pair of discrete variables given their
  marginal distributions is uniformed
  , i.e., returns the range of feasible bivariate
   correlations.
}
\examples{
Corrmatrix.limit(0,8,4)
}
