% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Simulate clinical score data}
\format{A data frame with 100 rows and 4 variables}
\usage{
data(dat)
}
\description{
A dataset that was generated using the function Generated.data
 function in the package.
}
\details{
The variables are as follows:

\itemize{
  \item V1 (0--8)
  \item V2 (0--8)
  \item V3 (0--8)
  \item V4 (0--8)
}
}
\keyword{datasets}
