% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF.R
\docType{data}
\name{DF}
\alias{DF}
\title{User defined correlation matrix}
\format{A data frame with 4 rows and 4 variables}
\usage{
data(DF)
}
\description{
A data frame that was used as correlation matrix to generate the dat dataset.
}
\details{
The variables are as follows:

\itemize{
  \item V1 (0--1)
  \item V2 (0--1)
  \item V3 (0--1)
  \item V4 (0--1)
}
}
\keyword{datasets}
