% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.data.R
\name{Generate.data}
\alias{Generate.data}
\title{Drawing a sample of discrete data}
\usage{
Generate.data(start, end, number.of.scores, corMatrix, x)
}
\arguments{
\item{start}{a number to set up the min value for the discrete variable.}

\item{end}{a number to set up the max value for the discrete variable.}

\item{number.of.scores}{a number represent  how many discrete variables.}

\item{corMatrix}{a  correlation matrix.}

\item{x}{the sample size.}
}
\value{
a \code{"x"} times \code{"number.of.scores"} matrix of values drawn
 from the \code{"number.of.scores"}-variate discrete
 r.v. with the uniform marginal distributions and desired correlation matrix
}
\description{
The function draws a sample from a multivariate discrete
variable with correlation matrix \code{"corMatrix"} and
uniformed marginal distributions with specific range from \code{"start"} to
\code{"end"}
}
\examples{
DDF<-matrix(as.double(data.matrix(DF)),4,4)
Generate.data(0,8, 4,DDF, 100)
}
