% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.power.R
\name{pca.power}
\alias{pca.power}
\title{Power caculation for pca}
\usage{
pca.power(dat, core = 7, beta = 0.5, sd = 5, iterations = 1000,
  seed = 17)
}
\arguments{
\item{dat}{a datafram that contains discrete data.}

\item{core}{a number to create cluster with desired number of cores}

\item{beta}{a effect size parameter for the mean of the simulated biomarkers}

\item{sd}{the standard deviation of the simulated biomarkers}

\item{iterations}{how many times you want to simulate the biomarkers}

\item{seed}{set the seed number for reproduce purpose}
}
\value{
a single number for the power of pca procedure
}
\description{
The function is use to calculate the pca power with desired times
of simulation \code{"iterations"}, effect size \code{"beta"}, standard
 deviation \code{"sd"}, and the target dataframe \code{"dat"}.
}
\examples{
pca.power(dat)
}
