% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_preserve_sum.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round a vector and make sure it adds up to a limit}
\source{
https://biostatmatt.com/archives/2902
}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{A vector that will rounded}

\item{digits}{The precision of the rounding. Set to 2 for probabilities to show as .15.}
}
\description{
This is useful for rounding a set of probabilities and making sure they sum to 1 or forcing percentages sum to 100.
}
